Move actual get logic into logic layer
