--- cloud.foundry.cli/src/main/java/cloud/foundry/cli/logic/LogicModule.java
@@ -1,9 +1,0 @@
-package cloud.foundry.cli.logic;
-
-/**
- * This class serves as a temporary dummy file. It makes sure that the otherwise empty package-folder is considered by
- * source control.
- * It has no functionality at the moment and might be renamed or deleted in the future.
- */
-public class LogicModule {
-}
